migrate courses logic from vuex to composable
