+++ kolibri/plugins/coach/frontend/modules/pluginModule.js
@@ -8,7 +8,6 @@ import exerciseDetail from './exerciseDetail';
 import groups from './groups';
 import lessonSummary from './lessonSummary';
 import lessonsRoot from './lessonsRoot';
-import coursesRoot from './coursesRoot';
 import classSummary from './classSummary';
 import coachNotifications from './coachNotifications';
 import questionDetail from './questionDetail';
@@ -153,7 +152,6 @@ export default {
   modules: {
     classSummary,
     coachNotifications,
-    coursesRoot,
     examReportDetail,
     exerciseDetail,
     groups,
